requirements.txt
absl-py==0.8.1
appdirs==1.4.3
argh==0.26.2
astor==0.7.1
attrs==19.3.0
backcall==0.1.0
bashplotlib==0.6.5
black==19.10b0
bokeh==1.4.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
configparser==4.0.2
cytoolz==0.10.0
dask==2.7.0
dask-glm==0.2.0
dask-jobqueue==0.7.0
dask-ml==1.0.0
decorator==4.4.1
distributed==2.7.0
docker-pycreds==0.4.0
entrypoints==0.3
flake8==3.7.9
fsspec==0.5.2
gast==0.2.2
gitdb2==2.0.6
GitPython==3.0.5
google-pasta==0.1.7
gql==0.2.0
graphql-core==1.1
grpcio==1.23.0
h5py==2.10.0
HeapDict==1.0.1
idna==2.8
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
line-profiler==2.1.1
llvmlite==0.30.0
locket==0.2.0
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
msgpack==0.6.2
multipledispatch==0.6.0
networkx==2.4
numba==0.46.0
numexpr==2.7.0
numpy==1.17.3
nvidia-ml-py==375.53.1
nvidia-ml-py3==7.352.0
olefile==0.46
opt-einsum==3.1.0
packaging==19.2
palettable==3.3.0
pandas==0.25.3
parso==0.5.1
partd==1.0.0
pathspec==0.6.0
pathtools==0.1.2
pbr==5.4.4
PeakUtils==1.3.2
periodictable==1.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
promise==2.3
prompt-toolkit==2.0.10
protobuf==3.9.2
psutil==5.6.5
ptyprocess==0.6.0
PubChemPy==1.0.4
pycairo==1.18.2
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.4.2
pynvml==8.0.4
pyparsing==2.4.4
pysmiles==1.0.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
regex==2019.8.19
requests==2.22.0
# Editable Git install with no remote (rotconml==0.1.0)
-e /data/sao/klee/projects/rotconml
scikit-learn==0.21.3
scipy==1.3.1
sentry-sdk==0.14.1
shortuuid==0.5.0
six==1.13.0
smmap2==2.0.5
sortedcontainers==2.1.0
subprocess32==3.5.4
tables==3.6.1
tblib==1.5.0
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.0
termcolor==1.1.0
toml==0.10.0
toolz==0.10.0
torch==1.3.1
torchvision==0.4.2
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.3
typed-ast==1.4.0
urllib3==1.25.8
wandb==0.8.22
watchdog==0.9.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
wrapt==1.11.2
xlrd==1.2.0
zict==1.0.0
